+++ xboot-fast/src/main/java/cn/exrick/xboot/config/properties/XbootIpLimitProperties.java
@@ -25,5 +25,5 @@ public class XbootIpLimitProperties {
     /**
      * 每单位时间内（毫秒）
      */
-    private Integer timeout = 1000;
+    private Long timeout = 1000L;
 }
