+++ xboot-module/xboot-core/src/main/java/cn/exrick/xboot/core/common/annotation/RateLimiter.java
@@ -12,5 +12,6 @@ import java.lang.annotation.*;
 public @interface RateLimiter {
 
     int limit() default 5;
-    int timeout() default 1000;
+
+    long timeout() default 1000;
 }
